fix: 필터 카테고리 적용 시 노트 개수 API DDD + 핵사고날로 리팩토링

+++ src/main/java/com/joojoo/api/filter/presentation/dto/response/FilterCountResponse.java
@@ -6,5 +6,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class FilterCountResponse {
-    private int filterCount;
+    private long filterCount;
 }
